@@ -1,71 +1,72 @@
 {
     "name": "open-capture-for-invoices",
-    "version": "2.2.4",
+    "version": "2.3.0",
     "scripts": {
+        "build-watch": "ng build --configuration development --watch",
+        "build-prod": "ng build --configuration production",
         "check-update": "ncu",
-        "build-watch": "ng build --watch --configuration development",
-        "build:prod": "NODE_ENV=production ng build --configuration production",
-        "reload-packages": "rm -Rf node_modules && rm -f package-lock.json && npm i",
+        "upgrade-all-packages-ncu": "ncu -u && npm run reload-packages",
         "extract-translations": "ngx-translate-extract --input ./src/frontend/ --output ./src/assets/i18n/frontend/*.json --clean --sort --format namespaced-json --marker _",
+        "reload-packages": "rm -Rf node_modules && rm -f package-lock.json && npm i --force",
         "lint": "ng lint"
     },
     "private": true,
     "dependencies": {
-        "@angular-material-extensions/select-country": "^6.0.1",
-        "@angular/animations": "~13.3.11",
-        "@angular/cdk": "^13.3.8",
-        "@angular/common": "~13.3.11",
-        "@angular/compiler": "~13.3.11",
-        "@angular/core": "~13.3.11",
-        "@angular/forms": "~13.3.11",
-        "@angular/material": "^13.3.8",
-        "@angular/material-moment-adapter": "^13.3.8",
-        "@angular/platform-browser": "~13.3.11",
-        "@angular/platform-browser-dynamic": "~13.3.11",
-        "@angular/router": "~13.3.11",
+        "@angular-material-extensions/select-country": "^7.0.0",
+        "@angular/animations": "~14.0.3",
+        "@angular/cdk": "^14.0.3",
+        "@angular/common": "~14.0.3",
+        "@angular/compiler": "~14.0.3",
+        "@angular/core": "~14.0.3",
+        "@angular/forms": "~14.0.3",
+        "@angular/material": "^14.0.3",
+        "@angular/material-moment-adapter": "^14.0.3",
+        "@angular/platform-browser": "~14.0.3",
+        "@angular/platform-browser-dynamic": "~14.0.3",
+        "@angular/router": "~14.0.3",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@fortawesome/fontawesome-free": "^6.1.1",
         "@ngx-translate/core": "^14.0.0",
         "@ngx-translate/http-loader": "^7.0.0",
-        "@schematics/angular": "^13.3.7",
+        "@schematics/angular": "^14.0.3",
         "@swimlane/ngx-charts": "^20.1.0",
-        "@syncfusion/ej2": "^20.1.60",
+        "@syncfusion/ej2": "^20.1.61",
         "jquery": "^3.6.0",
         "moment": "^2.29.3",
-        "ng-click-outside2": "^10.0.4",
+        "ng-click-outside2": "^11.1.0",
         "ng-sortgrid": "^6.0.1",
         "ng2-datepicker": "^12.0.0",
-        "ngx-file-drag-drop": "^5.1.3",
-        "ngx-mat-select-search": "^4.1.2",
+        "ngx-file-drag-drop": "^6.0.0",
+        "ngx-mat-select-search": "^4.2.0",
         "remove-accents": "^0.4.2",
         "rxjs": "~7.5.5",
         "svg-country-flags": "^1.2.10"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^13.3.7",
-        "@angular-devkit/build-webpack": "^0.1303.7",
-        "@angular-eslint/builder": "13.5.0",
-        "@angular-eslint/eslint-plugin": "13.5.0",
-        "@angular-eslint/eslint-plugin-template": "13.5.0",
-        "@angular-eslint/schematics": "13.5.0",
-        "@angular-eslint/template-parser": "13.5.0",
-        "@angular/cli": "~13.3.7",
-        "@angular/compiler-cli": "~13.3.11",
+        "@angular-devkit/build-angular": "^14.0.3",
+        "@angular-devkit/build-webpack": "^0.1400.3",
+        "@angular-eslint/builder": "14.0.0",
+        "@angular-eslint/eslint-plugin": "14.0.0",
+        "@angular-eslint/eslint-plugin-template": "14.0.0",
+        "@angular-eslint/schematics": "14.0.0",
+        "@angular-eslint/template-parser": "14.0.0",
+        "@angular/cli": "~14.0.3",
+        "@angular/compiler-cli": "~14.0.3",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
         "@types/d3-scale": "^4.0.2",
         "@types/d3-selection": "^3.0.2",
-        "@types/node": "^17.0.43",
-        "@typescript-eslint/eslint-plugin": "5.28.0",
-        "@typescript-eslint/parser": "5.28.0",
+        "@types/node": "^18.0.0",
+        "@typescript-eslint/eslint-plugin": "5.30.0",
+        "@typescript-eslint/parser": "5.30.0",
         "autoprefixer": "^10.4.7",
-        "eslint": "^8.17.0",
+        "eslint": "^8.18.0",
         "eslint-plugin-import": "latest",
         "eslint-plugin-jsdoc": "latest",
         "eslint-plugin-prefer-arrow": "latest",
         "postcss": "^8.4.14",
-        "tailwindcss": "^3.1.3",
+        "tailwindcss": "^3.1.4",
         "ts-node": "^10.8.1",
         "typescript": "~4.",
-        "zone.js": "~0.11.5"
+        "zone.js": "~0.11.6"
     }
 }
