@@ -133,7 +133,7 @@ export class SplitterFormListComponent implements OnInit {
                 confirmTitle        : this.translate.instant('GLOBAL.confirm'),
                 confirmText         : this.translate.instant('FORMS.confirm_duplicate', {"form": form}),
                 confirmButton       : this.translate.instant('GLOBAL.duplicate'),
-                confirmButtonColor  : "warn",
+                confirmButtonColor  : "green",
                 cancelButton        : this.translate.instant('GLOBAL.cancel'),
             },
             width: "600px",
@@ -173,7 +173,7 @@ export class SplitterFormListComponent implements OnInit {
                 confirmTitle        : this.translate.instant('GLOBAL.confirm'),
                 confirmText         : this.translate.instant('FORMS.confirm_enable', {"form": form}),
                 confirmButton       : this.translate.instant('GLOBAL.enable'),
-                confirmButtonColor  : "warn",
+                confirmButtonColor  : "green",
                 cancelButton        : this.translate.instant('GLOBAL.cancel'),
             },
             width: "600px",
